Updated to Lucene 7

+++ peppol-directory-indexer/src/test/java/com/helger/pd/indexer/lucene/PDLuceneTest.java
@@ -99,7 +99,7 @@ public final class PDLuceneTest
       if (aHits.length == 0)
         return null;
 
-      final int numTotalHits = results.totalHits;
+      final long numTotalHits = results.totalHits;
       assertEquals (1, numTotalHits);
 
       /*
